+++ packages/axway-cli-auth/src/auth/list.js
@@ -37,13 +37,14 @@ export default {
 		const check = process.platform === 'win32' ? '√' : '✔';
 		const now = Date.now();
 		const pretty = require('pretty-ms');
-		const table = createTable([ 'Account Name', 'Organization', 'Region', 'Type', 'Expires' ]);
+		const table = createTable([ 'Account Name', 'Organization', 'Team', 'Region', 'Type', 'Expires' ]);
 
-		for (const { default: def, auth, isPlatform, name, org } of accounts) {
+		for (const { default: def, auth, isPlatform, name, org, team } of accounts) {
 			const { access, refresh } = auth.expires;
 			table.push([
 				`${def ? green(`${check} ${name}`) : `  ${name}`}`,
 				!org || !org.name ? 'n/a' : org.id ? `${org.name} (${org.id})` : org.name,
+				team ? `${team.name} (${team.guid})` : 'n/a',
 				org?.region || 'US',
 				isPlatform ? 'Platform' : 'Service',
 				pretty((refresh || access) - now, { secDecimalDigits: 0, msDecimalDigits: 0 })