+++ packages/axway-cli-auth/src/auth/login.js
@@ -55,6 +55,7 @@ export default {
 	async action({ argv, cli, console }) {
 		const { default: snooplogg } = require('snooplogg');
 		const { getAuthConfigEnvSpecifier, initSDK, isHeadless } = require('@axway/amplify-cli-utils');
+		const { renderAccountInfo } = require('../lib/info');
 		const { prompt } = require('enquirer');
 
 		// prompt for the username and password
@@ -109,7 +110,7 @@ export default {
 		});
 		let account;
 		const authConfigEnvSpecifier = getAuthConfigEnvSpecifier(sdk.env.name);
-		const { highlight } = snooplogg.styles;
+		const { highlight, note } = snooplogg.styles;
 
 		// perform the login
 		const manual = !argv.launchBrowser;
@@ -151,11 +152,15 @@ export default {
 				if (argv.json) {
 					account.default = config.get(`${authConfigEnvSpecifier}.defaultAccount`) === account.name;
 					console.log(JSON.stringify(account, null, 2));
-				} else if (account.isPlatform && account.org?.name) {
-					console.log(`You are already logged into ${highlight(account.org.name)} as ${highlight(account.user.email || account.name)}.`);
+					return;
+				}
+
+				if (account.isPlatform && account.org?.name) {
+					console.log(`You are already logged into ${highlight(account.org.name)} ${note(`(${account.org.guid})`)} as ${highlight(account.user.email || account.name)}.`);
 				} else {
 					console.log(`You are already logged in as ${highlight(account.user.email || account.name)}.`);
 				}
+				console.log(await renderAccountInfo(account, config, sdk));
 				return;
 			}
 
@@ -183,12 +188,12 @@ export default {
 		}
 
 		if (account.isPlatform && account.org?.name) {
-			console.log(`You are logged into ${highlight(account.org.name)} as ${highlight(account.user.email || account.name)}.`);
+			console.log(`You are logged into ${highlight(account.org.name)} ${note(`(${account.org.guid})`)} as ${highlight(account.user.email || account.name)}.`);
 		} else {
 			console.log(`You are logged in as ${highlight(account.user.email || account.name)}.`);
 		}
 
-		console.log(`The current region is set to ${highlight(config.get('region', account.org?.region || 'US'))}.`);
+		console.log(await renderAccountInfo(account, config, sdk));
 
 		// set the current
 		if (accounts.length === 1 || account.default) {