+++ packages/axway-cli-auth/src/auth/switch.js
@@ -6,13 +6,15 @@ export default {
 			callback: ({ ctx, value }) => ctx.jsonMode = value,
 			desc: 'Disables prompting and outputs selected account and org as JSON'
 		},
-		'--org [guid|id|name]': 'The organization to switch to'
+		'--org [guid|id|name]': 'The org to switch to; required when --json and user belongs to more than one org',
+		'--team [guid|name]': 'The org team to select; required when --json and selected org has more than one team'
 	},
 	async action({ argv, cli, console }) {
 		const { default: snooplogg } = require('snooplogg');
 		const { getAuthConfigEnvSpecifier, initSDK, isHeadless } = require('@axway/amplify-cli-utils');
+		const { renderAccountInfo } = require('../lib/info');
 		const { prompt } = require('enquirer');
-		const { highlight } = snooplogg.styles;
+		const { highlight, note } = snooplogg.styles;
 		const { config, sdk } = initSDK({
 			baseUrl:  argv.baseUrl,
 			env:      argv.env,
@@ -73,11 +75,13 @@ export default {
 		account.default = true;
 		config.set(`${authConfigEnvSpecifier}.defaultAccount`, account.name);
 		config.delete(`${authConfigEnvSpecifier}.defaultOrg.${account.hash}`);
+		config.delete(`${authConfigEnvSpecifier}.defaultTeam.${account.hash}`);
 		config.save();
 
 		if (account.isPlatform) {
 			// determine the org
-			const selectedOrg = argv.org || (account?.hash && config.get(`${authConfigEnvSpecifier}.defaultOrg.${account.hash}`));
+			const defaultOrg = account?.hash && config.get(`${authConfigEnvSpecifier}.defaultOrg.${account.hash}`);
+			const selectedOrg = argv.org || defaultOrg;
 			const org = selectedOrg && account.orgs.find(o => o.guid === selectedOrg || o.id === selectedOrg || o.name === selectedOrg);
 
 			if (account.isPlatformTooling) {
@@ -98,7 +102,6 @@ export default {
 						throw new Error('Must specify --org when --json is set and the selected account has multiple organizations');
 					}
 
-					const defaultOrg = config.get(`${authConfigEnvSpecifier}.defaultOrg.${account.hash}`);
 					const choices = account.orgs
 						.map(org => {
 							org.toString = () => org.name;
@@ -147,7 +150,80 @@ export default {
 				});
 			}
 
+			if (account.org?.teams) {
+				// determine the team
+				const defaultTeam = account?.hash && config.get(`${authConfigEnvSpecifier}.defaultTeam.${account.hash}`);
+				const selectedTeam = String(argv.team || defaultTeam || '');
+				let team = selectedTeam && account.org.teams.find(t => t.guid.toLowerCase() === selectedTeam.toLowerCase() || t.name.toLowerCase() === selectedTeam.toLowerCase());
+
+				if (!team) {
+					if (argv.team) {
+						// if there was an explicit --org that wasn't found, then we error for tooling users as web doesn't care
+						const err = `Unable to find team "${argv.team}"`;
+						err.code = 'ERR_NOT_FOUND';
+						err.details = `Available teams:\n${account.org.teams.map(t => `  ${highlight(t.name)} ${note(`(${t.guid})`)}`).join('\n')}`;
+						throw err;
+					}
+
+					if (account.org.teams.length === 1) {
+						team = account.org.teams[0];
+					} else if (account.org.teams.length > 1) {
+						if (argv.json) {
+							throw new Error('Must specify --team when --json is set and the selected account has multiple teams');
+						}
+
+						const choices = account.org.teams
+							.map(team => {
+								team.toString = () => team.name;
+								return {
+									guid:    team.guid,
+									message: `${team.name} (${team.guid})`,
+									value:   team
+								};
+							})
+							.sort((a, b) => a.message.localeCompare(b.message));
+						const initial = choices.findIndex(team => team.guid === defaultTeam);
+						const { prompt } = require('enquirer');
+
+						team = (await prompt({
+							choices,
+							format: function () {
+								// for some reason, enquirer doesn't print the selected value using the primary
+								// (green) color for select prompts, so we just force it for all prompts
+								return this.style(this.value);
+							},
+							initial,
+							message: 'Select an team to use',
+							name: 'team',
+							styles: {
+								em(msg) {
+									// stylize emphasised text with just the primary color, no underline
+									return this.primary(msg);
+								}
+							},
+							type: 'select'
+						})).team;
+
+						console.log();
+					}
+				}
+
+				if (team) {
+					account.team = {
+						default: team.default,
+						guid:    team.guid,
+						name:    team.name,
+						roles:   team.users?.find(u => u.guid === account.user.guid)?.roles || [],
+						tags:    team.tags
+					};
+					await sdk.authClient.updateAccount(account);
+				}
+			}
+
 			config.set(`${authConfigEnvSpecifier}.defaultOrg.${account.hash}`, account.org.guid);
+			if (account.team) {
+				config.set(`${authConfigEnvSpecifier}.defaultTeam.${account.hash}`, account.team.guid);
+			}
 			config.save();
 		}
 
@@ -156,9 +232,11 @@ export default {
 		if (argv.json) {
 			console.log(JSON.stringify(account, null, 2));
 		} else if (account.isPlatform && account.org?.name) {
-			console.log(`Default account set to ${highlight(account.user.email || account.name)} in ${highlight(account.org.name)}`);
+			console.log(`Default account set to ${highlight(account.user.email || account.name)} in ${highlight(account.org.name)} ${note(`(${account.org.guid})`)}`);
 		} else {
 			console.log(`Default account set to ${highlight(account.user.email || account.name)}`);
 		}
+
+		console.log(await renderAccountInfo(account, config, sdk));
 	}
 };