+++ packages/axway-cli-auth/src/auth/whoami.js
@@ -13,13 +13,17 @@ export default {
 		}
 	},
 	async action({ argv, console }) {
-		const { initSDK } = await import('@axway/amplify-cli-utils');
+		const { getAuthConfigEnvSpecifier, initSDK } = require('@axway/amplify-cli-utils');
+		const { renderAccountInfo } = require('../lib/info');
 		const { config, sdk } = initSDK({
 			baseUrl:  argv.baseUrl,
 			env:      argv.env,
 			realm:    argv.realm
 		});
 		let accounts = await sdk.auth.list({ validate: true });
+		for (const account of accounts) {
+			account.default = account.name === config.get(`${getAuthConfigEnvSpecifier(account.auth.env)}.defaultAccount`);
+		}
 
 		if (argv.accountName) {
 			// eslint-disable-next-line security/detect-non-literal-regexp
@@ -31,20 +35,26 @@ export default {
 			console.log(JSON.stringify(accounts, null, 2));
 		} else {
 			const { default: snooplogg } = require('snooplogg');
-			const { highlight } = snooplogg.styles;
+			const { highlight, note } = snooplogg.styles;
+
 			if (accounts.length) {
-				let region = config.get('region');
+				let acct;
+
 				for (const account of accounts) {
 					if (account.isPlatform && account.org?.name) {
-						console.log(`You are logged into ${highlight(account.org.name)} as ${highlight(account.user.email || account.name)}.`);
+						console.log(`You are logged into ${highlight(account.org.name)} ${note(`(${account.org.guid})`)} as ${highlight(account.user.email || account.name)}.`);
 					} else {
 						console.log(`You are logged in as ${highlight(account.user.email || account.name)}.`);
 					}
-					if (!region && account.default) {
-						region = account.org?.region;
+					if (account.default) {
+						acct = account;
 					}
 				}
-				console.log(`The current region is set to ${highlight(region || 'US')}.`);
+
+				if (acct) {
+					acct = await sdk.auth.selectTeam(acct, config.get(`${getAuthConfigEnvSpecifier(acct.auth.env)}.defaultTeam.${acct.hash}`));
+					console.log(await renderAccountInfo(acct, config, sdk));
+				}
 			} else if (argv.accountName) {
 				console.log(`The account ${highlight(argv.accountName)} is not logged in.`);
 			} else {